# ENH: Vectorize argsort and argselect with AVX2