@lesteve 2024-02-01
There was a new failure in scikit-learn when testing against numpy dev because of this change, see https://github.com/scikit-learn/scikit-learn/pull/28326 for more details. The `np.argsort` changes enough that the number of clusters in HDBSCAN is 3 with numpy dev, and 2 with numpy 1.26.3.
 
The scikit-learn tests pass in 0a4b2b8 (previous merge commit in main) and fail in 221427b.

I could not find any mention of this change in the [changelog](https://numpy.org/devdocs/release/2.0.0-notes.html), but maybe I missed it?

I think this would be worth to adding a changelog entry about this change to indicate that np.argsort and np.argselect results may change in numpy 2.0.