@rgommers 2024-03-05
> I think this would be worth to adding a changelog entry about this change to indicate that np.argsort and np.argselect results may change in numpy 2.0.

I verified that this is indeed common when sorting integers.

```python
>>> import hashlib
>>> import numpy as np

>>> rng = np.random.RandomState(seed=123098)  # minor: reproducers across versions should use RandomState
>>> x = rng.randint(100, size=10_000)
>>> hashlib.sha256(np.argsort(x).tobytes()).hexdigest()
```

I'll include this note in my next release notes PR:
```
Minor changes in behavior of sorting functions
----------------------------------------------

Due to algorithmic changes and use of SIMD code, sorting functions with methods
that aren't stable may return slightly different results in 2.0.0 compared to
1.26.x. This includes the default method of `~numpy.sort` and `~numpy.argsort`.
```